+++ project/src/main/java/org/ironmaple/simulation/seasonspecific/reefscape2025/Arena2025Reefscape.java
@@ -7,6 +7,7 @@ import edu.wpi.first.math.geometry.Translation2d;
 import java.util.Arrays;
 import java.util.List;
 import org.ironmaple.simulation.SimulatedArena;
+import org.ironmaple.simulation.seasonspecific.reefscape2025.reef.ReefscapeReefSimulation;
 import org.ironmaple.utils.FieldMirroringUtils;
 
 /**
